Refactor: constants에 prize-contants 추가
